refactor: move ReportContent/OverpaymentContent to commonMain
Splits ReportContent and OverpaymentContent out of their wasmJsMain
Screen files into commonMain, moving RedemptionFormState alongside
OverpaymentContent since it's part of that composable's parameter
type. Enables PreviewScreenshotGeneratorTest on the JVM target for
both report screens.
