+++ feature/report/src/wasmJsMain/kotlin/feature/report/OverpaymentViewModel.kt
@@ -11,30 +11,8 @@ import core.auth.toJsString
 import core.network.MoneyRepository
 import core.network.ReportRepository
 import kotlinx.coroutines.launch
-import model.MonthlyMoney
-import model.MonthlyMoneyStatus
 import model.OverpaymentRedemptionRequest
 import model.UserBalance
-
-private const val DEFAULT_REDEMPTION_NOTE = "過払い金から支払い"
-
-data class RedemptionFormState(
-    val selectedUid: String = "",
-    val selectedYearMonth: String = "",
-    val amountText: String = "",
-    val noteText: String = DEFAULT_REDEMPTION_NOTE,
-    val isSaving: Boolean = false,
-    val error: String? = null,
-    val monthData: MonthlyMoney? = null,
-) {
-    val isMonthFrozen: Boolean get() = monthData?.status == MonthlyMoneyStatus.FROZEN
-    val canSubmit: Boolean
-        get() =
-            selectedUid.isNotEmpty() &&
-                (amountText.toLongOrNull() ?: 0L) > 0L &&
-                !isSaving &&
-                !isMonthFrozen
-}
 
 data class OverpaymentUiState(
     val balances: List<UserBalance> = emptyList(),
